Fix Experiment 1 to use bf16, matching the real GPU baseline precision
Was loading the model in fp16 with the deprecated torch_dtype= argument.
Switched to dtype=bfloat16 to match vLLM's precision on the real GPU
runs -- precision differences can flip close-call outputs, so this
keeps local experiments comparable to the real baseline.

+++ experiments/kv_contamination.py
@@ -153,7 +153,8 @@ def main():
 
     print(f"Loading {MODEL_NAME} (tokenizer + weights) ...")
     tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, torch_dtype=torch.float16)
+    # bf16, not fp16 -- matches the real GPU baseline run's precision.
+    model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, dtype=torch.bfloat16)
     model.to(device)
     model.eval()
     num_layers = model.config.num_hidden_layers
